Moved each sample to a separate folder. Added new publishing profile
