+++ SampleCommand/Program.cs
@@ -1,12 +1,16 @@
 ﻿namespace SampleCommand
 {
+    using SampleCommand.NamedSample;
+    using SampleCommand.ValuesSample;
+    using SampleCommand.VerbsSample;
+
     internal static class Program
     {
         internal static void Main(string[] args)
         {
-            ////new ValuesSampleCommand().Parse(args);
-            ////new NamedOptionsCommand().Parse(args);
-            ////new VerbsCommand().Parse(args);
+            //new ValuesSampleCommand().Parse(args);
+            //new NamedOptionsCommand().Parse(args);
+            //new VerbsCommand().Parse(args);
         }
     }
 }
