+++ SampleCommand/VerbsSample/ShowMessageVerb.cs
@@ -1,9 +1,9 @@
-﻿namespace SampleCommand
+﻿namespace SampleCommand.VerbsSample
 {
     using CommandLine;
 
-    [Verb("reverse", HelpText = "Reverses the message and shows the result.")]
-    internal class ReverseMessageVerb
+    [Verb("show", HelpText = "Shows the message.")]
+    internal class ShowMessageVerb
     {
         [Option('r', "repetitions", Required = true, HelpText = "Number of times the messages should be repeated.")]
         public int Repetitions { get; set; }
